Rename @branch/dtos to @branch/types
The package mostly contains kysely DB row types (closer to ORM models
than DTOs), so name it for what it holds: shared/types exporting
@branch/types.
