+++ apps/backend/lambdas/auth/package.json
@@ -10,7 +10,7 @@
     "package": "npm run build && cd dist && zip -r ../lambda.zip . -x '*.map' 'dev-server.*' 'swagger-utils.*'"
   },
   "devDependencies": {
-    "@branch/dtos": "file:../../../../shared/dtos",
+    "@branch/types": "file:../../../../shared/types",
     "@jest/globals": "^30.2.0",
     "@types/aws-lambda": "^8.10.131",
     "@types/jest": "^30.0.0",
