+++ apps/backend/lambdas/tools/lambda-cli.js
@@ -48,7 +48,7 @@ function templatePackageJson() {
     "package": "npm run build && cd dist && zip -r ../lambda.zip . -x '*.map' 'dev-server.*' 'swagger-utils.*'"
   },
   "devDependencies": {
-    "@branch/dtos": "file:../../../../shared/dtos",
+    "@branch/types": "file:../../../../shared/types",
     "@types/aws-lambda": "^8.10.131",
     "@types/node": "^20.11.30",
     "ts-node": "^10.9.2",
