--- shared/dtos/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@branch/dtos",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Shared DTO type definitions for Branch lambdas",
-  "types": "index.d.ts",
-  "files": [
-    "*.d.ts"
-  ]
-}
